subtask_2/requirements.txt
google-generativeai
torch
transformers
tqdm
scikit-learn
spacy
